fix: error handler fast refresh 이슈 해결
